Move slide actions to own file
