+++ lib/utils.js
@@ -1,6 +1,4 @@
 'use strict'
-const fs = require('fs')
-const path = require('path')
 
 exports.isSQLite = function isSQLite (sequelize) {
   return sequelize.options.dialect === 'sqlite'
@@ -26,11 +24,4 @@ exports.getImageMimeType = function getImageMimeType (imagePath) {
     default:
       return undefined
   }
-}
-
-exports.isRevealTheme = function isRevealTheme (theme) {
-  if (fs.existsSync(path.join(__dirname, '..', 'public', 'build', 'reveal.js', 'css', 'theme', theme + '.css'))) {
-    return theme
-  }
-  return undefined
 }
