+++ lib/web/note/router.js
@@ -9,6 +9,7 @@ const { markdownParser } = require('../utils')
 const router = module.exports = Router()
 
 const noteActions = require('./actions')
+const slide = require('./slide')
 
 // get new note
 router.get('/new', response.postNote)
@@ -21,9 +22,9 @@ router.get('/s/:shortid', response.showPublishNote)
 // publish note actions
 router.get('/s/:shortid/:action', response.publishNoteActions)
 // get publish slide
-router.get('/p/:shortid', response.showPublishSlide)
+router.get('/p/:shortid', slide.showPublishSlide)
 // publish slide actions
-router.get('/p/:shortid/:action', response.publishSlideActions)
+router.get('/p/:shortid/:action', slide.publishSlideActions)
 // get note by id
 router.get('/:noteId', response.showNote)
 // note actions
